ml-inference/requirements.txt
# ml-inference/requirements.txt

# Core ML Framework
tensorflow>=2.13.0
numpy>=1.24.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0

# Image Processing
Pillow>=10.0.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Utilities
pydantic>=2.0.0